Updating main tools and FusionEKF.cpp

+++ src/tools.cpp
@@ -47,26 +47,26 @@ MatrixXd Tools::CalculateJacobian(const VectorXd& x_state) {
     MatrixXd Hj(3,4);
 
     //recover state parameters
-    double px = x_state(0);
-    double py = x_state(1);
-    double vx = x_state(2);
-    double vy = x_state(3);
+    float px = x_state(0);
+    float py = x_state(1);
+    float vx = x_state(2);
+    float vy = x_state(3);
 
     //check division by zero
-    double denom = px*px + py*py;
+    float denom = px*px + py*py;
     assert(denom > 0);
-    double denom2 = sqrt(denom);
-    double denom32 = pow(denom, 3/2);
+    float denom2 = sqrt(denom);
+    float denom32 = pow(denom, 3/2);
 
     //compute the Jacobian matrix
-    double H11 = px/denom2;
-    double H12 = py/denom2;
-    double H21 = -py/denom;
-    double H22 = px/denom;
-    double H31 = py*(vx*py-vy*px)/denom32;
-    double H32 = px*(vy*px-vx*py)/denom32;
-    double H33 = px/denom2;
-    double H34 = py/denom2;
+    float H11 = px/denom2;
+    float H12 = py/denom2;
+    float H21 = -py/denom;
+    float H22 = px/denom;
+    float H31 = py*(vx*py-vy*px)/denom32;
+    float H32 = px*(vy*px-vx*py)/denom32;
+    float H33 = px/denom2;
+    float H34 = py/denom2;
 
     Hj << H11, H12,   0,   0,
           H21, H22,   0,   0,
